Refactors Searchbar With Suspense
