+++ src/app/(home)/games/components/gamesHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './gamesHeader'
